+++ gui/frontend/wailsjs/go/models.ts
@@ -58,6 +58,7 @@ export namespace devices {
 	    type?: string;
 	    notes?: string;
 	    ips: string[];
+	    owns_cidrs?: string[];
 	
 	    static createFrom(source: any = {}) {
 	        return new Device(source);
@@ -69,6 +70,7 @@ export namespace devices {
 	        this.type = source["type"];
 	        this.notes = source["notes"];
 	        this.ips = source["ips"];
+	        this.owns_cidrs = source["owns_cidrs"];
 	    }
 	}
 	export class Segment {